NamesStorage was moved to the package "common"
