+++ src/main/kotlin/common/NamesStorage.kt
@@ -1,4 +1,4 @@
-package sessions
+package common
 
 import network.Client
