+++ src/main/kotlin/sessions/WaitingForNameSession.kt
@@ -1,5 +1,6 @@
 package sessions
 
+import common.NamesStorage
 import data.Data
 import data.Exit
 import data.Name
